fix: 「戻る」ボタンを配置するページを変更
